requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
discord.py==2.0.1
frozenlist==1.4.1
google-api-core==2.19.2
google-api-python-client==2.93.0
google-auth==2.21.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.65.0
httplib2==0.22.0
idna==3.8
multidict==6.0.5
oauthlib==3.2.2
proto-plus==1.24.0
protobuf==5.28.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.4
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.20
yarl==1.9.7
